merge : feat/goodsLikeBE into back

See merge request s06-bigdata-rec-sub2/S06P22E202!213

+++ backend/backend-java/src/main/java/com/ssafy/api/service/GoodsService.java
@@ -25,7 +25,7 @@ public interface GoodsService {
 
     public UserCommentRes goodsCommentUpdate(UserCommentReq userCommentReq,int commentSeq);
 
-    public void goodsCommentDelete(long commentSeq);
+    public void goodsCommentDelete(int commentSeq);
 
     public MyLikeGoodsRes getMyGoodsList(int userId, Pageable pageable);
